Refactored controllers
