+++ src/main/java/com/backend/webproject/HelloWorld.java
@@ -13,19 +13,4 @@ public class HelloWorld {
 	public String Hello() {
 		return "home";
 	}
-	
-	@RequestMapping("/login")
-	public String Login() {
-		return "login";
-	}
-	
-	@RequestMapping("/login_admin")
-	public String LoginAdmin() {
-		return "login_admin";
-	}
-	
-	@RequestMapping("/sign_up")
-	public String SignUp() {
-		return "sign_up";
-	}
 }
