+++ src/main/java/com/backend/webproject/controllers/AdminController.java
@@ -1,0 +1,13 @@
+package com.backend.webproject.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminController {
+	@RequestMapping("/login")
+	public String Login() {
+		return "login_admin";
+	}
+}
