+++ src/main/resources/application.properties
@@ -1,3 +1,9 @@
 server.port=8081
 spring.mvc.view.prefix: /myviews/
 spring.mvc.view.suffix: .jsp
+
+spring.datasource.url=jdbc:oracle:thin:@localhost:1521:XE
+spring.datasource.username=system
+spring.datasource.password=system
+
+spring.datasource.driver-class-name=com.oracle.database.jdbc
