+++ src/main/webapp/myviews/home.jsp
@@ -1,7 +1,7 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <!DOCTYPE html>
-<html>
+<html xmlns:th="http://www.thymeleaf.org">
 	<head>
 		<title>Home</title>
 	</head>
@@ -9,13 +9,13 @@
 		<h1>Pages</h1>
 		
 		<div>
-			<a href="/login">login</a>
+			<a href="/user/login">login</a>
 		</div>
 		<div>
-			<a href="/login_admin">login admin</a>
+			<a href="/admin/login">login admin</a>
 		</div>
 		<div>
-			<a href="/sign_up">sign up</a>
+			<a href="/user/sign_up">sign up</a>
 		</div>
 	</body>
 </html>
